+++ pyloremgen/generator/pdf_utils/cover_page_template.py
